requirements.txt
Django==4.2.3
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
